requirements.txt
streamlit>=1.22.0
langchain>=0.0.267
langchain-community>=0.0.16
neo4j>=5.14.0
pymupdf>=1.23.7
mistralai>=0.0.7
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.18.0
python-dotenv>=1.0.0
huggingface-hub>=0.16.4
sentence-transformers>=2.2.2 